% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamonds.R
\docType{data}
\name{diamonds}
\alias{diamonds}
\title{Diamonds data}
\format{
A data frame with 53,940 rows and 10 variables:
\describe{
  \item{carat}{the carat}
  \item{cut}{the cut}
  \item{color}{the color}
  \item{clarity}{how clear it is}
  \item{depth}{depth of the diamond}
  \item{table}{width at top of diamond}
  \item{price}{price of diamond}
  \item{x}{length in mm}
  \item{y}{width in mm}
  \item{z}{depth in mm}

}
}
\source{
https://ggplot2.tidyverse.org/reference/diamonds.html
}
\usage{
diamonds
}
\description{
The diamonds dataset
}
\references{
https://ggplot2.tidyverse.org/reference/diamonds.html
}
\keyword{datasets}
