% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onevar.R
\name{onevar}
\alias{onevar}
\title{Determines and provides the most suitable plot for a given variable}
\usage{
onevar(df, var, askente = TRUE)
}
\arguments{
\item{df}{The data frame or tibble}

\item{var}{The variable you wish to explore}

\item{askente}{Do you want Ente to choose the plot theme?
Default=TRUE}
}
\value{
a ggplot object
}
\description{
The function will automatically determine whether the given variables should be categorical or numerical.
Variables with more than 10 categories will be treated as numerical.
}
\examples{
onevar(diamonds, cut)
}
