% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threevar.R
\name{threevar}
\alias{threevar}
\title{Determines and provides the most suitable plot(s) for three given variables}
\usage{
threevar(df, var1, var2, var3, askente = TRUE)
}
\arguments{
\item{df}{The dataframe or tibble}

\item{var1}{First variable (numerical or categorical)}

\item{var2}{Second variable (numerical or categorical)}

\item{var3}{Third variable (numerical or categorical)}

\item{askente}{Do you want Ente to choose the plot theme?
Default=TRUE}
}
\value{
a ggplot object
}
\description{
The function will automatically determine whether the given variables should be categorical or numerical.
Variables with more than 10 categories will be treated as numerical. Note: barplots are for total counts whenever barplots are returned.
}
\examples{
threevar(diamonds, table, color, cut)
}
