% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twovar.R
\name{twovar}
\alias{twovar}
\title{Determines and provides the most suitable plot(s) for two given variables}
\usage{
twovar(df, var1, var2, askente = TRUE)
}
\arguments{
\item{df}{The dataframe or tibble}

\item{var1}{First variable}

\item{var2}{Second variable}

\item{askente}{Do you want Ente to choose the plot theme?
Default=TRUE}
}
\value{
a ggplot object
}
\description{
Determines and provides the most suitable plot(s) for two given variables
}
\examples{
twovar(diamonds, cut, price)
}
