% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threevar.R
\name{threevar}
\alias{threevar}
\title{Determines and provides the most suitable plot(s) for three given variables}
\usage{
threevar(df, var1, var2, var3, askente = TRUE)
}
\arguments{
\item{df}{The dataframe or tibble}

\item{var1}{First variable}

\item{var2}{Second variable}

\item{var3}{Third variable}

\item{askente}{Do you want Ente to choose the plot theme?
Default=TRUE}
}
\value{
a ggplot object
}
\description{
Determines and provides the most suitable plot(s) for three given variables
}
\examples{
threevar(diamonds, table, color, cut)
}
